requirements.txt
six==1.10.0
pyglet==1.2.4
cocos2d==0.6.4
